feat(search-page): move search bar component into separate file
