+++ src/components/pages/search/index.js
@@ -1,40 +1,22 @@
 /* eslint-disable no-restricted-syntax */
-import React, { useEffect } from 'react'
-import NProgress from 'nprogress'
+import React from 'react'
 import withSearch from '~components/utils/with-search'
 import SearchNoResults from '~components/search/search-no-results'
 import SearchResultSection from '~components/search/search-result-section'
+
+import Search from './search'
 import searchStyle from './search.module.scss'
-import searchIcon from '~images/icons/search-inverted.svg'
 
 import {
   types,
   useSearch,
-  querySearch,
   getHighlightResultOrExcerpt,
   getSanitizedSlug,
 } from '~context/search-context'
 
-const Search = withSearch(({ navigate, search }) => {
+const SearchAndResults = withSearch(({ navigate, search }) => {
   const [searchState, searchDispatch] = useSearch()
   const { query, results } = searchState
-
-  function setQuery(value) {
-    NProgress.start()
-    return searchDispatch({ type: 'setQuery', payload: value })
-  }
-
-  useEffect(() => {
-    if (search.q) {
-      setQuery(search.q)
-    }
-  }, [])
-
-  useEffect(() => {
-    if (query) {
-      querySearch(searchState, searchDispatch)
-    }
-  }, [query])
 
   const totalHits =
     query &&
@@ -42,50 +24,15 @@ const Search = withSearch(({ navigate, search }) => {
       (results[types.BLOG_POST].nbHits || 0) +
       (results[types.PAGE].nbHits || 0)
 
-  let searchEvent
-
   return (
     <>
-      <form
-        className={searchStyle.searchForm}
-        onSubmit={event => {
-          event.preventDefault()
-        }}
-      >
-        <button
-          type="button"
-          className={searchStyle.searchSubmit}
-          aria-label="Submit search"
-          onClick={() => query && navigate(`/search?q=${query}`)}
-        >
-          <img
-            src={searchIcon}
-            className={searchStyle.searchIcon}
-            alt=""
-            aria-hidden="true"
-          />
-        </button>
-        {/* eslint-disable jsx-a11y/label-has-associated-control */}
-        <label htmlFor="search-page-input" className={searchStyle.label}>
-          Search
-        </label>
-        <input
-          type="search"
-          id="search-page-input"
-          name="search"
-          autoComplete="off"
-          defaultValue={query || ''}
-          onChange={event => {
-            clearTimeout(searchEvent)
-            const { value } = event.currentTarget
-            searchEvent = setTimeout(() => {
-              setQuery(value)
-              window.history.pushState('', '', `?q=${value}`)
-            }, 300)
-          }}
-        />
-        {/* eslint-enable jsx-a11y/label-has-associated-control */}
-      </form>
+      <Search
+        query={query}
+        search={search}
+        navigate={navigate}
+        searchDispatch={searchDispatch}
+        searchState={searchState}
+      />
       {totalHits > 0 ? (
         <div className={searchStyle.searchResults}>
           <h2>
@@ -149,4 +96,4 @@ const Search = withSearch(({ navigate, search }) => {
   )
 })
 
-export default Search
+export default SearchAndResults
